+++ src/modules/auth/serializers/auth.serialize.ts
@@ -1,12 +1,6 @@
 import { User } from "../../user/entities/user.entity"
-import minio from "../../../core/helpers/minio"
 
 export class AuthSerializer {
-    private static async resolvePhotoUrl(photo?: string | null): Promise<string | null> {
-        if (!photo) return null
-        return await minio.getPresignedUrl(photo)
-    }
-
     static single(user: User) {
         return {
             id: user.id,
@@ -15,5 +9,9 @@ export class AuthSerializer {
             email: user.email,
             isActive: user.isActive,
         }
+    }
+
+    static collection(users: User[]) {
+        return users.map(u => this.single(u))
     }
 }