+++ src/modules/auth/validators/auth.validator.ts
@@ -34,8 +34,8 @@ export const RefreshTokenValidator = z.object({
 
 export type RefreshTokenValidator = z.infer<typeof RefreshTokenValidator>
 
-export const GoogleLoginSchema = z.object({
+export const GoogleLoginValidator = z.object({
     code: z.string().min(1, 'Code is required'),
 })
 
-export type GoogleLoginValidator = z.infer<typeof GoogleLoginSchema>
+export type GoogleLoginValidator = z.infer<typeof GoogleLoginValidator>