+++ src/modules/contact/interfaces/contact.repository.interface.ts
@@ -1,10 +1,6 @@
-import { EntityManager } from "typeorm"
 import { Contact } from "../entities/contact.entity"
+import { IBaseRepository } from "../../../core/interfaces/base.repository.interface"
 
-export interface IContactRepository {
+export interface IContactRepository extends IBaseRepository<Contact> {
     findAll(page: number, limit: number, q: string): Promise<{ data: Contact[]; total: number }>
-    findById(id: number): Promise<Contact | null>
-    save(data: Partial<Contact>, manager?: EntityManager): Promise<Contact>
-    merge(entity: Contact, data: Partial<Contact>): Contact
-    delete(id: number): Promise<void>
 }